Pre-Check update
The scaling governor check is based on a driver. In my case, when I switched to using uefi this driver wasn't being used by default and thus this check was constantly failing. Might want to do more about this in the future (performance setting check). For now, just pass the check if the file doesn't exist.